move(encoding): variables from unstable_base32.ts to _common32.ts
This is to allow unstable_base32_stream.ts to use it without exposing them to the public
